# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ClubOS Internal API Documentation
  version: 1.0.0
  description: |
    **Internal Developer Documentation for ClubOS API**

    This documentation covers ALL ClubOS APIs including public, member, admin, officer, and internal endpoints.
    It is intended for developers, tech leads, reviewers, and future maintainers.

    ## Authentication

    ClubOS uses session-based authentication with HttpOnly cookies. For local development and testing:

    - **Cookie-based auth**: The primary method. Session cookies are set after login.
    - **Test tokens (dev only)**: Use headers like `x-admin-test-token: dev-admin-token` for E2E tests.
    - **Bearer tokens (dev only)**: Format: `Bearer test-{role}-{memberId}` (e.g., `test-admin-123`)

    **IMPORTANT**: Never embed real tokens in documentation or code. Use placeholder values.

    ## Testing Endpoints Safely

    1. Use the dev environment only for testing
    2. Use test tokens with `x-admin-test-token` header
    3. Never test destructive operations against production data

    ## Error Responses

    All endpoints return standardized error responses:
    - `400` - Validation error (bad request body/params)
    - `401` - Unauthorized (missing or invalid auth)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Resource not found
    - `429` - Rate limited
    - `500` - Internal server error

  contact:
    name: ClubOS Development Team
  license:
    name: Proprietary
...
servers:
...
paths:
...
  /api/v1/events/{id}/status:
    get:
      tags: [Events]
      summary: Get event status
      description: Get current status and availability for an event.
      operationId: getEventStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event status
        '404':
          description: Event not found
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: clubos_session
      description: |
        Session cookie set after authentication.
        In development, you can also use `clubos_dev_session`.

    testToken:
      type: apiKey
      in: header
      name: x-admin-test-token
      description: |
        Test token for E2E testing (development only).
        Default: `dev-admin-token`
        Set via ADMIN_E2E_TOKEN environment variable.

    cronSecret:
      type: apiKey
      in: header
      name: x-cron-secret
      description: Secret token for cron job authentication.
...